chore: move components folder
